requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
numpy==1.24.3
pandas==2.1.1
openai==1.3.0
python-dotenv==1.0.0
pydantic==2.4.2
httpx==0.24.1 